Put the `IsSession` class in its own module
